Rename: oodns => resolver/ooniresolver
Yak shaving.
